default statis added
