+++ src/components/tabbedpane/TabbedPane.jsx
@@ -11,7 +11,7 @@ import BGPTable from "../../components/tabbedpane/bgp/bgpTable";
 import StpDataTable from "./stp/stpDataTable";
 
 import { useParams } from "react-router-dom";
-import { getAllDevicesURL, getStateURL } from "../../utils/backend_rest_urls";
+import { getAllDevicesURL } from "../../utils/backend_rest_urls";
 import PortGroupTable from "./portgroup/portGroupTable";
 import VlanTable from "./vlan/vlanTable";
 import "../../pages/home/home.scss";
@@ -32,7 +32,6 @@ const TabbedPane = () => {
   );
   const [dropdownOptions, setDropdownOptions] = useState([]);
   const [undoChanges, setUndoChanges] = useState(false);
-  const [orcaState, setOrcaState] = useState("NO CONFIGURATION");
 
   const setUpdateLog = useStoreLogs((state) => state.setUpdateLog);
 
@@ -41,7 +40,6 @@ const TabbedPane = () => {
       settabvalue(0);
       secureLocalStorage.setItem("selectedTab", tabvalue);
     }
-    getOrcaState();
 
     instance(getAllDevicesURL())
       .then((res) => {
@@ -56,24 +54,10 @@ const TabbedPane = () => {
   const handleTabs = (event, val) => {
     settabvalue(val);
     secureLocalStorage.setItem("selectedTab", val);
-    getOrcaState();
     setUpdateLog(true);
   };
 
-  const getOrcaState = () => {
-    instance(getStateURL(deviceIP))
-      .then((res) => {
-        console.log(res.data)
-        if (res.data) {
-            console.log('1')
-          setOrcaState(res.data.state);
-        } else {
-            console.log('2')
-          setOrcaState("NO CONFIGURATION");
-        }
-      })
-      .catch((err) => console.error(err));
-  };
+
 
   const onUndo = (event) => {
     setUndoChanges(true);
@@ -158,7 +142,7 @@ const TabbedPane = () => {
               selectedDeviceIp={deviceIP}
               height={height}
               reset={() => setUndoChanges(false)}
-              orcaState={orcaState}
+            //   orcaState={orcaState}
             />
           </div>
         )}
